+++ django_modern_rest/validation/endpint_metadata.py
@@ -1,6 +1,6 @@
 import dataclasses
 import inspect
-from collections.abc import Callable, Mapping, Set
+from collections.abc import Callable, Set
 from http import HTTPMethod, HTTPStatus
 from types import NoneType
 from typing import (
@@ -8,15 +8,12 @@ from typing import (
     Any,
     ClassVar,
     NewType,
-    TypeAlias,
     assert_never,
     cast,
 )
 
 from django.http import HttpResponse
 
-from django_modern_rest.cookies import NewCookie
-from django_modern_rest.errors import AsyncErrorHandlerT, SyncErrorHandlerT
 from django_modern_rest.exceptions import (
     EndpointMetadataError,
 )
@@ -40,57 +37,15 @@ from django_modern_rest.types import (
     is_safe_subclass,
     parse_return_annotation,
 )
+from django_modern_rest.validation.payload import (
+    ModifyEndpointPayload,
+    PayloadT,
+    ValidateEndpointPayload,
+)
 
 if TYPE_CHECKING:
     from django_modern_rest.controller import Blueprint, Controller
-    from django_modern_rest.openapi.objects import (
-        Callback,
-        ExternalDocumentation,
-        Reference,
-        SecurityRequirement,
-        Server,
-    )
 
-
-@dataclasses.dataclass(slots=True, frozen=True, kw_only=True, init=False)
-class _BasePayload:
-    # OpenAPI stuff:
-    summary: str | None = None
-    description: str | None = None
-    tags: list[str] | None = None
-    operation_id: str | None = None
-    deprecated: bool = False
-    security: list['SecurityRequirement'] | None = None
-    external_docs: 'ExternalDocumentation | None' = None
-    callbacks: 'dict[str, Callback | Reference] | None' = None
-    servers: list['Server'] | None = None
-
-    # Common fields:
-    validate_responses: bool | None = None
-    error_handler: SyncErrorHandlerT | AsyncErrorHandlerT | None = None
-    allow_custom_http_methods: bool = False
-    no_validate_http_spec: Set[HttpSpec] | None = None
-
-
-@dataclasses.dataclass(slots=True, frozen=True, kw_only=True)
-class ValidateEndpointPayload(_BasePayload):
-    """Payload created by ``@validate``."""
-
-    responses: list[ResponseSpec]
-
-
-@dataclasses.dataclass(slots=True, frozen=True, kw_only=True)
-class ModifyEndpointPayload(_BasePayload):
-    """Payload created by ``@modify``."""
-
-    status_code: HTTPStatus | None
-    headers: Mapping[str, NewHeader] | None
-    cookies: Mapping[str, NewCookie] | None
-    responses: list[ResponseSpec] | None
-
-
-#: Alias for different payload types:
-PayloadT: TypeAlias = ValidateEndpointPayload | ModifyEndpointPayload | None
 
 #: NewType for better typing safety, don't forget to resolve all responses
 #: before passing them to validation.
